+++ src/java/main/Dataset.java
@@ -3,6 +3,6 @@ package main;
 public interface Dataset {
     public BitSet[] getInitialCuts();
 
-    public int[] getCutCosts();
+    public double[] getCutCosts();
 
 }
